+++ client/package.json
@@ -17,7 +17,6 @@
     "redux": "^4.2.0",
     "redux-thunk": "^2.3.0",
     "web-vitals": "^2.1.4",
-    "axios": "^0.27.2",
     "react-slick": "^0.29.0",
     "slick-carousel": "^1.8.1"
   },
