+++ client/src/components/UserLogin/UserLogin.jsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useHistory, Link } from "react-router-dom";
+import { getUser } from "../../redux/StoreUsers/usersActions.js"
 import { useAuth0 } from "@auth0/auth0-react";
 import axios from "axios";
-import { getUser } from "../../redux/booksActions";
 
 export default function UserLogin() {
   const history = useHistory();
