+++ client/src/redux/StoreUsers/usersSlice.js
@@ -1,0 +1,16 @@
+import { createSlice } from "@reduxjs/toolkit";
+
+export const usersSlice = createSlice({
+  name: "users",
+  initialState: {
+    users: [],
+  },
+  reducers: {
+    getAllUsers: (state, action) => {
+      state.users = action.payload;
+    },
+  },
+});
+
+export const { getAllUsers } = usersSlice.actions;
+export default usersSlice.reducer;
