+++ client/src/redux/store.js
@@ -1,8 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit';
-import booksReducer from './booksSlice.js';
+import booksReducer from './StoreBooks/booksSlice.js';
+import userReducer  from './StoreUsers/usersSlice';
 
 export default configureStore({
   reducer: {
     books: booksReducer,
+    user: userReducer
   },
 });
